edited file structure to be better organized.
